Signatur änderbar Grid anpassbar Notenlängen nach Grid anpassbar Playhead nicht mehr von Takt zu Takt, sondern flüssig.

+++ src/main/java/fh/swf/NoteEvent.java
@@ -5,13 +5,13 @@ import lombok.Setter;
 
 @Setter @Getter
 public class NoteEvent {
-    private int column;
+    private double column;
     private int row;
     private int midiNote;
     private double length;
     private int channel;
 
-    public NoteEvent(int column, int row, double length, int channel) {
+    public NoteEvent(double column, int row, double length, int channel) {
         this.column = column;
         this.row = row;
         this.midiNote = 107 - row;
